requirements.txt
streamlit==1.31.0
pandas
plotly
sentence-transformers
pypdf
torch
requests
python-dotenv